requirements.txt
openai>=1.30.0
pydantic>=2.0.0
aiohttp>=3.9.0
Pillow>=10.0.0
numpy>=1.26.0
fastapi>=0.111.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
